% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STICCC_utilities.R
\name{computeDist}
\alias{computeDist}
\title{Compute pairwise distance between single cells in PCA space}
\usage{
computeDist(
  sce,
  est = TRUE,
  useGenes = FALSE,
  reduction = "PCA",
  nComponents = 10
)
}
\arguments{
\item{sce}{viccc object. Must have already executed the function runPCA().}

\item{est}{logical. Whether to estimate the max distance with a downsampled subset of data. Default TRUE.}

\item{useGenes}{logical. Whether to use full gene space to calculate distance. Setting this to TRUE
will significantly slow down the operation. Default FALSE.}

\item{reduction}{character. Which reduced space to use for distance calculation. Currently only supports "PCA"/}

\item{nComponents}{numeric. How many PCs to use for distance computation. Larger values will take
longer, especially for large datasets, but also incorporate more information. Default 10.}
}
\value{
viccc object with defined grid points in PCA space
}
\description{
Computes grid points for future steps that will smooth the inferred velocities
in PCA space.
}
