% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STICCC_utilities.R
\name{sticSE}
\alias{sticSE}
\title{Create VICCC object.}
\usage{
sticSE(
  topo,
  exprMat = NA,
  normData,
  topoName,
  expName = NA,
  radius = 0.05,
  minNeighbors = 5,
  scalingFactor = 1,
  gridScalingFactor = 1,
  verbose = T,
  useOriginalFeatures = F,
  nPCs = NA,
  plotDim = "PCA"
)
}
\arguments{
\item{topo}{data.frame. Topology table.}

\item{exprMat}{data.frame or matrix. Expression matrix with genes as rows and cells as columns.}

\item{normData}{data.frame or matrix. Normalized (log2+1) expression matrix.}

\item{topoName}{character. Name of the topology, used in output file names.}

\item{expName}{character. Name for the experiment (if you intend to do multiple analyses with the same topology).
Default <topoName>_SCE.}

\item{radius}{numeric. Proportion between 0 and 1 of the maximum pairwise distance to evaluate when computing velocity
of a cell. Default 0.15}

\item{minNeighbors}{numeric. Minimum number of neighbors within specified radius to compute a velocity for a cell. Cells
with fewer than this number of neighbors will be ignored, but may be counted as neighbors for other eligible cells.
Default 5.}

\item{scalingFactor}{numeric. Factor to multiply the magnitude of vectors for easier visual interpretation. Default 1.}

\item{gridScalingFactor}{numeric. Factor to multiply the magnitude of vectors when applying grid-based smoothing. Default 1.}

\item{verbose}{logical. Whether to print progress statements during analysis with this VICCC object. Default TRUE.}

\item{useOriginalFeatures}{logical. Whether to calculate vectors in gene space instead of projected space. Default FALSE.}

\item{nPCs}{numeric. How many components to use for distance calculation and vector output.}

\item{plotDim}{character. What reduced dimension to calculate and display vectors in. Can be "PCA", "UMAP", or "TSNE". Default "PCA".}
}
\value{
SingleCellExperiment object modified slightly for VICCC methods.
}
\description{
Returns a SingleCellExperiment object initialized for use with VICCC methods.
}
